Add install from source option to build.sh
Update build.sh to support install from source scenario.

+++ src/linux/Packaging.Linux/build.sh
@@ -19,7 +19,6 @@ make_absolute () {
 # Building
 #####################################################################
 echo "Building Packaging.Linux..."
-
 # Parse script arguments
 for i in "$@"
 do
@@ -30,6 +29,10 @@ case "$i" in
     ;;
     --version=*)
     VERSION="${i#*=}"
+    shift # past argument=value
+    ;;
+    --install-from-source=*)
+    INSTALL_FROM_SOURCE=${i#*=}
     shift # past argument=value
     ;;
     *)
@@ -59,22 +62,28 @@ if [ -z "$VERSION" ]; then
     die "--version was not set"
 fi
 
-ARCH="`dpkg-architecture -q DEB_HOST_ARCH`"
-if test -z "$ARCH"; then
-  die "Could not determine host architecture!"
+if [ $INSTALL_FROM_SOURCE = false ]; then
+    ARCH="`dpkg-architecture -q DEB_HOST_ARCH`"
+    if test -z "$ARCH"; then
+    die "Could not determine host architecture!"
+    fi
 fi
 
 # Outputs
 PAYLOAD="$PROJ_OUT/payload/$CONFIGURATION"
 SYMBOLOUT="$PROJ_OUT/payload.sym/$CONFIGURATION"
 
-TAROUT="$PROJ_OUT/tar/$CONFIGURATION"
-TARBALL="$TAROUT/gcmcore-linux_$ARCH.$VERSION.tar.gz"
-SYMTARBALL="$TAROUT/symbols-linux_$ARCH.$VERSION.tar.gz"
+if [ $INSTALL_FROM_SOURCE = false ]; then
+    TAROUT="$PROJ_OUT/tar/$CONFIGURATION"
+    TARBALL="$TAROUT/gcmcore-linux_$ARCH.$VERSION.tar.gz"
+    SYMTARBALL="$TAROUT/symbols-linux_$ARCH.$VERSION.tar.gz"
 
-DEBOUT="$PROJ_OUT/deb/$CONFIGURATION"
-DEBROOT="$DEBOUT/root"
-DEBPKG="$DEBOUT/gcmcore-linux_$ARCH.$VERSION.deb"
+    DEBOUT="$PROJ_OUT/deb/$CONFIGURATION"
+    DEBROOT="$DEBOUT/root"
+    DEBPKG="$DEBOUT/gcmcore-linux_$ARCH.$VERSION.deb"
+else
+    INSTALL_LOCATION="/usr/local"
+fi
 
 # Cleanup payload directory
 if [ -d "$PAYLOAD" ]; then
@@ -89,7 +98,13 @@ if [ -d "$SYMBOLOUT" ]; then
 fi
 
 # Ensure directories exists
-mkdir -p "$PAYLOAD" "$SYMBOLOUT" "$DEBROOT"
+mkdir -p "$PAYLOAD" "$SYMBOLOUT"
+
+if [ $INSTALL_FROM_SOURCE = false ]; then
+    mkdir -p "$DEBROOT"
+else
+    mkdir -p "$INSTALL_LOCATION"
+fi
 
 # Publish core application executables
 echo "Publishing core application..."
@@ -135,40 +150,44 @@ mv "$PAYLOAD"/*.pdb "$SYMBOLOUT" || exit 1
 echo "Build complete."
 
 #####################################################################
-# PACKING
+# PACKING AND INSTALLING
 #####################################################################
-echo "Packing Packaging.Linux..."
-# Cleanup any old archive files
-if [ -e "$TAROUT" ]; then
-    echo "Deleteing old archive '$TAROUT'..."
-    rm "$TAROUT"
-fi
-
-# Ensure the parent directory for the archive exists
-mkdir -p "$TAROUT" || exit 1
-
 # Set full read, write, execute permissions for owner and just read and execute permissions for group and other
 echo "Setting file permissions..."
 /bin/chmod -R 755 "$PAYLOAD" || exit 1
 
-# Build binaries tarball
-echo "Building binaries tarball..."
-pushd "$PAYLOAD"
-tar -czvf "$TARBALL" * || exit 1
-popd
+if [ $INSTALL_FROM_SOURCE = false ]; then
+    echo "Packing Packaging.Linux..."
+    # Cleanup any old archive files
+    if [ -e "$TAROUT" ]; then
+        echo "Deleteing old archive '$TAROUT'..."
+        rm "$TAROUT"
+    fi
 
-# Build symbols tarball
-echo "Building symbols tarball..."
-pushd "$SYMBOLOUT"
-tar -czvf "$SYMTARBALL" * || exit 1
-popd
+    # Ensure the parent directory for the archive exists
+    mkdir -p "$TAROUT" || exit 1
 
-# Build .deb
-INSTALL_TO="$DEBROOT/usr/local/share/gcm-core/"
-LINK_TO="$DEBROOT/usr/local/bin/"
-mkdir -p "$DEBROOT/DEBIAN" "$INSTALL_TO" "$LINK_TO" || exit 1
+    # Build binaries tarball
+    echo "Building binaries tarball..."
+    pushd "$PAYLOAD"
+    tar -czvf "$TARBALL" * || exit 1
+    popd
+
+    # Build symbols tarball
+    echo "Building symbols tarball..."
+    pushd "$SYMBOLOUT"
+    tar -czvf "$SYMTARBALL" * || exit 1
+    popd
+
+    # Build .deb
+    INSTALL_TO="$DEBROOT/usr/local/share/gcm-core/"
+    LINK_TO="$DEBROOT/usr/bin/"
+    mkdir -p "$DEBROOT/DEBIAN" "$INSTALL_TO" "$LINK_TO" || exit 1
 
 # make the debian control file
+# this is purposefully not indented, see
+# https://stackoverflow.com/questions/9349616/bash-eof-in-if-statement
+# for details
 cat >"$DEBROOT/DEBIAN/control" <<EOF
 Package: gcmcore
 Version: $VERSION
@@ -183,13 +202,25 @@ Description: Cross Platform Git Credential Manager command line utility.
  For more information see https://aka.ms/gcm
 EOF
 
+    dpkg-deb --build "$DEBROOT" "$DEBPKG" || exit 1
+else
+    echo "Installing..."
+
+    # Install directories
+    INSTALL_TO="$INSTALL_LOCATION/share/gcm-core/"
+    LINK_TO="$INSTALL_LOCATION/bin/"
+    MESSAGE="Install complete."
+fi
+
+mkdir -p "$INSTALL_TO" "$LINK_TO"
+
 # Copy all binaries and shared libraries to target installation location
 cp -R "$PAYLOAD"/* "$INSTALL_TO" || exit 1
 
 # Create symlink
-ln -s -r "$INSTALL_TO/git-credential-manager-core" \
-    "$LINK_TO/git-credential-manager-core" || exit 1
+if [ ! -f "$LINK_TO/git-credential-manager-core" ]; then
+    ln -s -r "$INSTALL_TO/git-credential-manager-core" \
+        "$LINK_TO/git-credential-manager-core" || exit 1
+fi
 
-dpkg-deb --build "$DEBROOT" "$DEBPKG" || exit 1
-
-echo "Pack complete."
+echo $MESSAGE
